pyproject.toml
[project]
name = "generate-news-article"
version = "0.1.0"
description = "Anthropic APIでMarkdown記事を生成するCLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.95.0",
    "fire>=0.7.1",
    "google-genai>=1.73.1",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
generate-news-article = "main:run"
